Retrieve rdf-toolkit
This repository is planned to be a submodule of UCO.  To prevent a
circular import, UCO will need to be removed as a submodule of this
repository.

This patch relieves this repository's reliance on UCO for rdf-toolkit,
instead copying the Makefile and hash as they appear in UCO 1.2.0.  Make
dependencies are also updated to note what requires the UCO Git
submodule to have been initiated.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>