Change shape type to address multiple-imports issue
If this shapes graph happens to be imported twice---from a lattice of
imports where X imports this, Y imports this, and Z imports X and
Y---the `sh:inversePath` causes a SHACL parser to fail.  `sh:path`
cannot have two values, and each instance of the `sh:inversePath` node
will always appear to be its own value, because SHACL syntactically
requires the subject of `sh:inversePath` to be a blank node.

Put another way with explicit illustration: because the
`sh:PropertyShape`, as spelled before this patch, is identified as an
IRI, the OWL shapes graph being imported twice "collapses" most of the
triples in the `PropertyShape`'s definition, but cannot collapse
`sh:path`:

```turtle
# This ...
sh-owl:incompatibleWith-imported-version-iri-shape
	a sh:PropertyShape ;
	sh:path [
		sh:inversePath owl:versionIRI ;
	] ;
	.
# ...combined with this ...
sh-owl:incompatibleWith-imported-version-iri-shape
	a sh:PropertyShape ;
	sh:path [
		sh:inversePath owl:versionIRI ;
	] ;
	.
# ...yields this:
sh-owl:incompatibleWith-imported-version-iri-shape
	a sh:PropertyShape ;
	sh:path
		[
			sh:inversePath owl:versionIRI ;
		] ,
		[
			sh:inversePath owl:versionIRI ;
		]
		;
	.
```

There are two potential resolutions to this issue:

1. Require users of this graph to uniquely import each ontology IRI.
2. Turn the property shape into a blank node inherent to a node shape.
3. Define a procedure for merging SHACL shape graphs that relies on
   graph isomorphism to recognize when blank nodes are redundant with
   one another, collapsing them.

The drawback to resolution 1 is a significant code-logistics issue for
all adopters.

The drawback to resolution 2 is SHACL runtime increases by a linear
factor with each import of the OWL shapes, because the blank node
attached with `sh:property` will be copied for each import.

Resolution 3 has not been tested.

This patch applies resolution 2, leaving benefit available for adopters
able to implement resolution 1, but not requiring adopters implement
resolution 1.

No effects were observed on Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>